Cas valeur négative > 2yx31

+++ mts/src/main/java/com/devoteam/srit/xmlloader/diameter/dictionary/Dictionary.java
@@ -205,7 +205,7 @@ public class Dictionary
     
     public Application getApplication(String key)
     {
-        if(Utils.isInteger(key))  return getApplicationById(Integer.parseInt(key));
+        if(Utils.isInteger(key))  return getApplicationById((int)Long.parseLong(key));
         else                      return getApplicationByName(key);
     }
